9/requirement.txt
bert-serving-client==1.9.1
bert-serving-server==1.9.1
pandas>=0.24.2
numpy>=1.16.4
